requirements.txt
easydict==1.9
numpy==1.18.5
opencv-python==4.3.0.36
pandas==1.0.5
pandocfilters==1.4.2
Pillow @ file:///tmp/build/80754af9/pillow_1594307295532/work
prefetch-generator==1.0.1
preshed==3.0.2
pretrainedmodels==0.7.4
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.12.2
python-dateutil==2.8.1
pytorch-warmup==0.0.4
scipy==1.5.1
seaborn==0.10.1
tensorboardX==2.1
torch==1.4.0
torchvision==0.5.0
visdom==0.1.8.9
